replaces class named utils by class named helper + merge BdvHandleHelper in class BdvUtils renamed BdvHandleHelper
