+++ pom.xml.releaseBackup
@@ -93,7 +93,7 @@
 		<connection>scm:git:git://github.com/bigdataviewer/bigdataviewer-playground</connection>
 		<developerConnection>scm:git:git@github.com:bigdataviewer/bigdataviewer-playground</developerConnection>
 		<url>https://github.com/bigdataviewer/bigdataviewer-playground/</url>
-    <tag>HEAD</tag>
+    <tag>bigdataviewer-playground-0.2.7</tag>
   </scm>
 
 	<issueManagement>
