+++ src/main/java/sc/fiji/bdvpg/scijava/services/SourceAndConverterBdvDisplayService.java
@@ -35,20 +35,16 @@ import net.imglib2.converter.Converter;
 import net.imglib2.util.Pair;
 import org.scijava.command.CommandService;
 import org.scijava.object.ObjectService;
-import org.scijava.options.OptionsService;
 import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
-import org.scijava.prefs.PrefService;
 import org.scijava.script.ScriptService;
 import org.scijava.service.AbstractService;
 import org.scijava.service.SciJavaService;
 import org.scijava.service.Service;
 import sc.fiji.bdvpg.bdv.projector.Projection;
 import sc.fiji.bdvpg.scijava.command.bdv.BdvWindowCreatorCommand;
-import sc.fiji.bdvpg.scijava.services.ui.SourceFilterNode;
-import sc.fiji.bdvpg.scijava.services.ui.SpimDataFilterNode;
 import sc.fiji.bdvpg.services.SourceAndConverterServices;
-import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterUtils;
+import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterHelper;
 
 import java.lang.ref.WeakReference;
 import java.util.*;
@@ -273,7 +269,7 @@ public class SourceAndConverterBdvDisplayService extends AbstractService impleme
 
         // If no ConverterSetup is built then build it
         if ( bdvSourceAndConverterService.sacToMetadata.getIfPresent(sac).get( CONVERTER_SETUP ) == null) {
-            ConverterSetup setup = SourceAndConverterUtils.createConverterSetup(sac);
+            ConverterSetup setup = SourceAndConverterHelper.createConverterSetup(sac);
             bdvSourceAndConverterService.sacToMetadata.getIfPresent(sac).put( CONVERTER_SETUP,  setup );
         }
