+++ src/main/java/sc/fiji/bdvpg/sourceandconverter/importer/SourceAndConverterDuplicator.java
@@ -30,7 +30,7 @@ package sc.fiji.bdvpg.sourceandconverter.importer;
 
 import bdv.viewer.SourceAndConverter;
 import sc.fiji.bdvpg.services.SourceAndConverterServices;
-import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterUtils;
+import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterHelper;
 
 import java.util.function.Function;
 
@@ -57,14 +57,14 @@ import java.util.function.Function;
         if (sourceAndConverter.asVolatile() != null) {
             sac = new SourceAndConverter(
                     sourceAndConverter.getSpimSource(),
-                    SourceAndConverterUtils.cloneConverter(sourceAndConverter.getConverter(), sourceAndConverter),
+                    SourceAndConverterHelper.cloneConverter(sourceAndConverter.getConverter(), sourceAndConverter),
                     new SourceAndConverter(sourceAndConverter.asVolatile().getSpimSource(),
-                            SourceAndConverterUtils.cloneConverter(sourceAndConverter.asVolatile().getConverter(), sourceAndConverter.asVolatile()))
+                            SourceAndConverterHelper.cloneConverter(sourceAndConverter.asVolatile().getConverter(), sourceAndConverter.asVolatile()))
             );
         } else {
             sac = new SourceAndConverter(
                     sourceAndConverter.getSpimSource(),
-                    SourceAndConverterUtils.cloneConverter(sourceAndConverter.getConverter(), sourceAndConverter));
+                    SourceAndConverterHelper.cloneConverter(sourceAndConverter.getConverter(), sourceAndConverter));
         }
         SourceAndConverterServices.getSourceAndConverterService().register(sac);
         return sac;
