+++ src/main/java/sc/fiji/bdvpg/sourceandconverter/importer/Wave3DSourceGetter.java
@@ -32,7 +32,7 @@ import bdv.util.Procedural3DImageShort;
 import bdv.viewer.Source;
 import bdv.viewer.SourceAndConverter;
 import net.imglib2.FinalInterval;
-import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterUtils;
+import sc.fiji.bdvpg.sourceandconverter.SourceAndConverterHelper;
 
 import java.util.function.Supplier;
 
@@ -49,6 +49,6 @@ public class Wave3DSourceGetter implements Runnable, Supplier<SourceAndConverter
                         p -> (int) ((Math.sin(p[0]/20)*Math.sin(p[1]/40)*Math.sin(p[2]/5)+1)*100)
                 ).getSource(new FinalInterval(new long[]{0,0,0}, new long[]{512,512,512}),"Wave 3D");
 
-        return SourceAndConverterUtils.createSourceAndConverter(s);
+        return SourceAndConverterHelper.createSourceAndConverter(s);
     }
 }
